requirements.txt
django==1.8
django-ckeditor==5.0.3
